feat(component): support budget amount
there is a new amount prop for budget, this change make able users to add a initial amount when
creating a budget. I also update the ui to display this value, calculate budget data using this new
prop and also add a new progress bar component to display how left/over the budger is.